+++ src/components/element/promoCodeCart.ts
@@ -1,4 +1,4 @@
-import { IPromoCode } from "../../basic";
+import { IPromoCode } from "../../type";
 import { LocalStorageManager } from "../controller/localStorage";
 import { elementGenerator } from "../controller/taggenerator";
 
@@ -35,7 +35,9 @@ export class PromoCodeCart {
     const promoTest = elementGenerator.createParagraph({text: 'Promo for test: "rs", "epm"', className: 'example-promo'});
     this._inputPromo.addEventListener('input', (event) => this.checkInputPromo(event));
     fragment.append(this._newPriceHTML, ...applyPromo, this._inputPromo, promoTest);
+
     window.addEventListener('storage', () => this.updateTitlePrice());
+
     return fragment;
   }
